Handle entity spawning packest

+++ Sources/Core/Sources/Network/Protocol/Packets/Play/Clientbound/EntityHeadLookPacket.swift
@@ -4,7 +4,7 @@ public struct EntityHeadLookPacket: ClientboundPacket {
   public static let id: Int = 0x3b
   
   public var entityId: Int
-  public var headYaw: UInt8
+  public var headYaw: Float
 
   public init(from packetReader: inout PacketReader) throws {
     entityId = packetReader.readVarInt()
